Break up MappingTools module into separate files
