+++ heidrun/mapping_tools.rb
@@ -2,227 +2,5 @@
 # Static methods for commonly-performed evaluations and assignments
 #
 module MappingTools
-
-  ##
-  # Static methods for evaluations and assignments that can be shared between
-  # mappings of MARC providers.
-  #
-  module MARC
-
-    # Lambdas for `.select` calls performed on record nodes (which are
-    # Krikri::XmlParser::Value objects)
-    #
-    # @example:
-    #   cf_007 = r.node.children
-    #             .select(&MappingTools::MARC::IS_CF7_NODE)
-    #             .first.children.to_s
-    #
-    IS_LEADER_NODE = lambda { |node| node.name == 'leader' }
-    IS_CF7_NODE = lambda { |node|
-      node.name == 'controlfield' && node[:tag] == '007'
-    }
-    IS_CF8_NODE = lambda { |node|
-      node.name == 'controlfield' && node[:tag] == '008'
-    }
-
-    module_function
-
-    ##
-    # Given options representing leader and controlfield values, return an
-    # array of applicable genre (edm:hasType) controlled-vocabulary terms
-    #
-    # Options are as follows, and are all Strings:
-    #   - leader:  value of MARC 'leader' element
-    #   - cf_007:  value of MARC 'controlfield' with 'tag' attribute '007'
-    #   - cf_008:  value of MARC 'controlfield' with 'tag' attribute '008'
-    #   - cf_970a:  TODO
-    #
-    # @param opts [Hash] Options, as outlined above
-    # @return [Array]
-    def genre(opts)
-      genres = []
-      args = [genres, opts]
-      # TODO: insert handling of optional cf_970a for Hathi, as the first
-      # evaluation.
-      assign_language(*args) || assign_musical_score(*args) ||
-        assign_manuscript(*args) || assign_maps(*args) ||
-        assign_projected(*args) || assign_two_d(*args) ||
-        assign_nonmusical_sound(*args) || assign_musical_sound(*args)
-      genres << 'Government Document' if government_document?(opts[:cf_008])
-      genres
-    end
-
-    def assign_language(genres, opts)
-      if language_material?(opts[:leader])
-        if monograph?(opts[:leader])
-          genres << 'Book'
-        elsif serial?(opts[:leader])
-          if newspapers?(opts[:cf_008])
-            genres << 'Newspapers'
-          else
-            genres << 'Serial'
-          end
-        elsif mono_component_part?(opts[:leader])
-          genres << 'Book'
-        else
-          genres << 'Serial'
-        end
-        true
-      else
-        false
-      end
-    end
-
-    def assign_musical_score(genres, opts)
-      if notated_music?(opts[:leader]) || manu_notated_music?(opts[:leader])
-        genres << 'Musical Score'
-        true
-      else
-        false
-      end
-    end
-
-    def assign_manuscript(genres, opts)
-      if manu_lang_material?(opts[:leader])
-        genres << 'Manuscript'
-        true
-      else
-        false
-      end
-    end
-
-    def assign_maps(genres, opts)
-      if cart_material?(opts[:leader]) || manu_cart_material?(opts[:leader])
-        genres << 'Maps'
-        true
-      else
-        false
-      end
-    end
-
-    # projected media
-    def assign_projected(genres, opts)
-      if projected_medium?(opts[:leader])
-        if slide?(opts[:cf_007]) || transparency?(opts[:cf_007])
-          genres << 'Photograph / Pictorial Works'
-          true
-        elsif film_video?(opts[:cf_007])
-          genres << 'Film / Video'
-          true
-        else
-          false
-        end
-      else
-        false
-      end
-    end
-
-    # two-dimensional nonprojectable graphic
-    def assign_two_d(genres, opts)
-      if two_d_nonproj_graphic?(opts[:leader])
-        genres << 'Photograph / Pictorial Works'
-        true
-      else
-        false
-      end
-    end
-
-    def assign_nonmusical_sound(genres, opts)
-      if nonmusical_sound?(opts[:leader])
-        genres << 'Nonmusic Audio'
-        true
-      else
-        false
-      end
-    end
-
-    def assign_musical_sound(genres, opts)
-      if musical_sound?(opts[:leader])
-        genres << 'Music'
-        true
-      else
-        false
-      end
-    end
-
-    def language_material?(s)
-      s[6] == 'a'
-    end
-
-    def monograph?(s)
-      s[7] == 'm'
-    end
-
-    def newspapers?(s)
-      s[21] == 'n'
-    end
-
-    def serial?(s)
-      s[7] == 's'
-    end
-
-    # monographic component part
-    def mono_component_part?(s)
-      s[7] == 'a'
-    end
-
-    def notated_music?(s)
-      s[6] == 'c'
-    end
-
-    # manuscript notated music
-    def manu_notated_music?(s)
-      s[6] == 'd'
-    end
-
-    # manuscript language material
-    def manu_lang_material?(s)
-      s[6] == 't'
-    end
-
-    # cartographic material
-    def cart_material?(s)
-      s[6] == 'e'
-    end
-
-    # manuscript cartographic material
-    def manu_cart_material?(s)
-      s[6] == 'f'
-    end
-
-    def projected_medium?(s)
-      s[6] == 'g'
-    end
-
-    def slide?(s)
-      s[1] == 's'
-    end
-
-    def transparency?(s)
-      s[1] == 't'
-    end
-
-    def film_video?(s)
-      %w(c d f o).include?(s[1])
-    end
-
-    # two-dimensional non-projectable graphic
-    def two_d_nonproj_graphic?(s)
-      s[6] == 'k'
-    end
-
-    # nonmusical sound recording
-    def nonmusical_sound?(s)
-      s[6] == 'i'
-    end
-
-    # musical sound recording
-    def musical_sound?(s)
-      s[6] == 'j'
-    end
-
-    def government_document?(s)
-      %w(a c f i l m o s).include?(s[28])
-    end
-  end
+  require_relative 'mapping_tools/marc'
 end
